==> nbs/01_model_building/gtex/02_gtex_plier_vs_ora.ipynb <==
# GTEx model with different priors

## Load libraries

if (!requireNamespace("PLIER", quietly = TRUE)) {
    devtools::install_github("wgmao/PLIER")
}

# 3. Install PLIER2 (mchikina/PLIER2) if not already installed
if (!requireNamespace("PLIER2", quietly = TRUE)) {
    REPO_PATH <- "/home/msubirana/Documents/pivlab/PLIER2"  # adjust
    remotes::install_local(REPO_PATH, force = TRUE, dependencies = FALSE)
}

library(bigstatsr)
library(data.table)
library(dplyr)
library(rsvd)
library(glmnet)
library(Matrix)
library(knitr)
library(here)
library(PLIER2)
library(clusterProfiler)
library(org.Hs.eg.db)
library(dplyr)


source(here("config.R"))

## Output directory

output_data_dir <- config$GTEx$DATASET_FOLDER
dir.create(output_data_dir, showWarnings = FALSE, recursive = TRUE)

options(repr.matrix.max.rows = Inf)

gtex_plier2_kegg <- readRDS(file.path(output_data_dir, "gtex_KEGG_2021_Human_PLIER2.rds"))

kegg_plier2_summary <- gtex_plier2_kegg$summary %>%
    dplyr::filter(FDR < 0.05 & AUC > 0.7) %>%
    dplyr::group_by(LV) %>%
    dplyr::summarise(pathway = paste(pathway, collapse = ', '), .groups = "drop") %>%
    dplyr::arrange(as.numeric(gsub("[^0-9]", "", LV)))

gtex_baseRes <- readRDS(file.path(output_data_dir, "gtex_PLIER2_baseRes.rds"))

gtex_genes <- readRDS(file.path(output_data_dir, "gtex_genes.rds"))

colnames(gtex_baseRes$Z) <- paste0('LV', seq_len(ncol(gtex_baseRes$Z)))
rownames(gtex_baseRes$Z) <- gtex_genes

head(gtex_baseRes$Z)

kegg_plier2_Z <- data.frame(gtex_baseRes$Z)

Z <- gtex_baseRes$Z

map <- bitr(rownames(Z), fromType = "SYMBOL", toType = "ENTREZID", OrgDb = org.Hs.eg.db)
map <- map[!duplicated(map$SYMBOL), ]

lv_names <- colnames(Z)

top_syms <- setNames(vector("list", length(lv_names)), lv_names)

head(top_syms$LV1)

kegg_plier2_Z %>% 
dplyr::select(LV1)  %>% 
arrange(desc(LV1)) %>%
tail(10)

# for (lv in lv_names) {
#   v <- Z[, lv]
#   idx <- which(!is.na(v) & v > 0)
#   if (length(idx) == 0L) { top_syms[[lv]] <- character(0); next }
#   ord <- idx[order(v[idx], decreasing = TRUE)]
#   top_syms[[lv]] <- rownames(Z)[ord]
# }

# top_entrez <- lapply(top_syms, function(syms) {
#   ids <- map$ENTREZID[match(syms, map$SYMBOL)]
#   unique(as.character(ids[!is.na(ids)]))
# })

# kegg_list <- lapply(top_entrez, function(ids) {
#   if (!length(ids)) return(NULL)
#   enrichKEGG(gene = ids, organism = "hsa", qvalueCutoff = 0.05)
# })

for (lv in lv_names) {
  v <- Z[, lv]
  ord <- order(v, decreasing = TRUE, na.last = NA)
  n_top <- max(1L, ceiling(length(ord) * 0.05))
  top_syms[[lv]] <- rownames(Z)[ord][seq_len(n_top)]
}

top_entrez <- lapply(top_syms, function(syms) {
  ids <- map$ENTREZID[match(syms, map$SYMBOL)]
  unique(as.character(ids[!is.na(ids)]))
})

kegg_list <- lapply(top_entrez, function(ids) {
  if (!length(ids)) return(NULL)
  enrichKEGG(gene = ids, organism = "hsa", qvalueCutoff = 0.05)
})

kegg_list$LV15@result  %>% 
dplyr::filter(p.adjust < 0.05)  %>% 
dplyr::pull(Description) 

kegg_plier2_summary  %>%  dplyr::filter(LV == "LV15")

kegg_list$LV102@result  %>% 
dplyr::filter(p.adjust < 0.05)  %>% 
dplyr::pull(Description) 

kegg_plier2_summary  %>%  dplyr::filter(LV == "LV102")

kegg_list$LV12@result  %>% 
dplyr::filter(p.adjust < 0.05)  %>% 
dplyr::pull(Description) 

kegg_plier2_summary  %>%  dplyr::filter(LV == "LV12")

kegg_list$LV117@result  %>% 
dplyr::filter(p.adjust < 0.05)  %>% 
dplyr::pull(Description) 

kegg_plier2_summary  %>%  dplyr::filter(LV == "LV117")

kegg_list$LV23@result  %>% 
dplyr::filter(p.adjust < 0.05)  %>% 
dplyr::pull(Description) 

kegg_plier2_summary  %>%  dplyr::filter(LV == "LV23")

kegg_list$LV44@result  %>% 
dplyr::filter(p.adjust < 0.05)  %>% 
dplyr::pull(Description) 

kegg_plier2_summary  %>%  dplyr::filter(LV == "LV44")
